Combine the SHA3 family into a single module
Merge the KeccakF, Keccak_sponge and SHAKE modules into the SHA3
module. There is no compelling reason to expose the Keccak-f[1600]
permutation or the Keccak[c] sponge directly in the API at this point,
so the only useful entry points are SHA3 and SHAKE. The amount of code
that is used by one of SHA3 and SHAKE but not both is very small, so
any code saving from including both when you only need one may well be
negated by the overhead of doing a cross-module call for the
underlying Keccak functions.

There is now a single sha3 module exposing both mbedtls_sha3_xxx() and
mbedtls_shake_xxx(). The Keccak functions are not exposed directly.
This module is controlled by the symbol MBEDTLS_SHA3_C and allows an
alternative implementation by defining MBEDTLS_SHA3_ALT, following the
library's usual conventions.

The full list of changes in this commit are as follows:

* Move all code from `keccacf.c`, `keccak_sponge.c` and `shake.c` to `sha3.c`.
* Remove the unused function `mbedtls_keccakf_clone`.
* Declare all Keccak functions as `static`.
* Move declarations from `shake.h` to `sha3.h`.
* Move macro definitions and function documentations from `keccacf.h` and `keccak_sponge.h` to `sha3.c`.
* Move type definitions from `keccacf.h` and `keccak_sponge.h` to `sha3.h`.
* Merge each error `MBEDTLS_ERR_SHAKE_xxx`, `MBEDTLS_ERR_KECCACKF_xxx` and `MBEDTLS_ERR_KECCAK_SPONGE_xxx` with `MBEDTLS_ERR_SHA3_xxx`. Replace `MBEDTLS_ERR_KECCAK_SPONGE_NOT_SETUP` with `MBEDTLS_ERR_SHA3_BAD_NOT_STARTED`.
* The conversion function `mbedtls_convert_sponge_result` is now the identity function and will be simplified away in a future commit.
* Remove the `keccakf`, `keccak_sponge` and `shake` modules and update `config.h`, `error.h`, `generate_errors.pl` and build scripts accordingly.
* Update the description of `MBEDTLS_SHA3_C` in `config.h` to include SHAKE.
* Rename `MBEDTLS_KECCAKF_STATE_SIZE_xxx` to `KECCAKF_STATE_SIZE_xxx` to satisfy `check-names.sh`.
* Run `scripts/generate_errors.pl`, `scripts/generate_features.pl` and `scripts/generate_visualc_files.pl`.
