+++ include/mbedtls/config.h
@@ -317,8 +317,6 @@
 //#define MBEDTLS_DHM_ALT
 //#define MBEDTLS_ECJPAKE_ALT
 //#define MBEDTLS_GCM_ALT
-//#define MBEDTLS_KECCAKF_ALT
-//#define MBEDTLS_KECCAK_SPONGE_ALT
 //#define MBEDTLS_MD2_ALT
 //#define MBEDTLS_MD4_ALT
 //#define MBEDTLS_MD5_ALT
@@ -330,7 +328,6 @@
 //#define MBEDTLS_SHA256_ALT
 //#define MBEDTLS_SHA3_ALT
 //#define MBEDTLS_SHA512_ALT
-//#define MBEDTLS_SHAKE_ALT
 //#define MBEDTLS_XTEA_ALT
 
 /*
@@ -2655,58 +2652,16 @@
 /**
  * \def MBEDTLS_SHA3_C
  *
- * Enable the SHA-3 cryptographic hash algorithms.
+ * Enable the SHA-3 cryptographic hash algorithms and the associated
+ * family of extendable-output functions (XOF) SHAKE.
  *
  * Module:  library/sha3.c
  * Caller:  library/md.c
  *
- * Requires: MBEDTLS_KECCAKF_C, MBEDTLS_KECCAK_SPONGE_C
- *
- * This module adds support for SHA3-224, SHA3-256, SHA3-384, and SHA3-512.
+ * This module adds support for SHA3-224, SHA3-256, SHA3-384, SHA3-512,
+ * SHAKE128 and SHAKE256.
  */
 #define MBEDTLS_SHA3_C
-
-/**
- * \def MBEDTLS_SHAKE_C
- *
- * Enable the SHA-3 eXtensible Output Functions (XOF).
- *
- * Module:  library/shake.c
- * Caller:
- *
- * Requires: MBEDTLS_KECCAKF_C, MBEDTLS_KECCAK_SPONGE_C
- *
- * This module adds support for SHAKE128 and SHAKE256.
- */
-#define MBEDTLS_SHAKE_C
-
-/**
- * \def MBEDTLS_KECCAKF_C
- *
- * Enable the SHA-3 cryptographic hash algorithms.
- *
- * Module:  library/keccakf.c
- * Caller:  library/keccak_sponge.c
- *
- * This module adds support for the Keccak-f[1600] permutation.
- * This module is required by for the Keccak Sponge module.
- */
-#define MBEDTLS_KECCAKF_C
-
-/**
- * \def MBEDTLS_KECCAKF_C
- *
- * Enable the Sponge construction based on Keccak-f[1600].
- *
- * Module:  library/keccak_sponge.c
- * Caller:  library/sha3.c
- *
- * Requires: MBEDTKS_KECCAKF_C
- *
- * This module adds support for the sponge construction based on Keccak-f[1600].
- * This module is required by for the SHA-3 module.
- */
-#define MBEDTLS_KECCAK_SPONGE_C
 
 /**
  * \def MBEDTLS_SSL_CACHE_C
