+++ include/mbedtls/error.h
@@ -70,8 +70,6 @@
  * PBKDF2    1  0x007C-0x007C
  * HMAC_DRBG 4                  0x0003-0x0009
  * CCM       3                  0x000D-0x0011
- * Keccak-f  1                  0x0015-0x0015
- * Sponge    3                  0x001B-0x001F
  * ARC4      1                  0x0019-0x0019
  * MD2       1                  0x002B-0x002B
  * MD4       1                  0x002D-0x002D
@@ -83,7 +81,6 @@
  * CHACHA20  3                  0x0051-0x0055
  * POLY1305  3                  0x0057-0x005B
  * SHA-3     3                  0x0061-0x0065
- * SHAKE     3                  0x0067-0x006B
  *
  * High-level module nr (3 bits - 0x0...-0x7...)
  * Name      ID  Nr of Errors
