+++ include/mbedtls/sha3.h
@@ -2,7 +2,9 @@
  * \file sha3.h
  *
  * \brief SHA-3 cryptographic hash functions
-          (SHA3-224, SHA3-256, SHA3-384, SHA3-512)
+ *        (SHA3-224, SHA3-256, SHA3-384, SHA3-512)
+ *        and associated extendable-output functions (XOF)
+ *        (SHAKE128, SHAKE256).
  *
  * Reference: National Institute of Standards and Technology (NIST).
  * _SHA-3 Standard: Permutation-Based Hash and Extendable-Output Functions._
@@ -37,7 +39,7 @@
 #include MBEDTLS_CONFIG_FILE
 #endif
 
-#include "keccak_sponge.h"
+#include <stddef.h>
 
 #ifdef __cplusplus
 extern "C" {
@@ -59,8 +61,49 @@ typedef enum
 }
 mbedtls_sha3_type_t;
 
+/**
+ * \brief Designators for algorithms in the SHAKE family.
+ */
+typedef enum
+{
+    MBEDTLS_SHAKE128, /**< SHAKE128 */
+    MBEDTLS_SHAKE256  /**< SHAKE256 */
+}
+mbedtls_shake_type_t;
+
 #if !defined(MBEDTLS_SHA3_ALT)
 // Regular implementation
+
+#include <stdint.h>
+
+/* The context structure for Keccak-f[1600].
+ * This structure may change or may be removed in future versions of
+ * the library and may not be present in alternative implementations of
+ * this module.
+ */
+typedef struct
+{
+    uint64_t state[5][5];
+    uint64_t temp[5][5];
+}
+mbedtls_keccakf_context;
+
+/* The context structure for the Keccak[c] sponge.
+ * This structure may change or may be removed in future versions of
+ * the library and may not be present in alternative implementations of
+ * this module.
+ */
+typedef struct
+{
+    mbedtls_keccakf_context keccakf_ctx;
+    unsigned char queue[1600 / 8]; /* store partial block data (absorbing) or pending output data (squeezing) */
+    size_t queue_len;              /* queue length (in bits) */
+    size_t rate;                   /* sponge rate (in bits) */
+    size_t suffix_len;             /* length of the suffix (in bits) (range 0..8) */
+    int state;                     /* Current state (absorbing/ready to squeeze/squeezing) */
+    unsigned char suffix;          /* suffix bits appended to message, before padding */
+}
+mbedtls_keccak_sponge_context;
 
 /**
  * \brief               The context structure for SHA-3 operations.
@@ -78,6 +121,22 @@ typedef struct
     size_t block_size;  /* block size in bytes */
 }
 mbedtls_sha3_context;
+
+/**
+ * \brief               The context structure for SHAKE operations.
+ *
+ * \note                This structure may change in future versions of the
+ *                      library. Hardware-accelerated implementations may
+ *                      use different structures. Therefore applications
+ *                      should not access the context directly, but instead
+ *                      should use the functions in this module.
+ */
+typedef struct
+{
+    mbedtls_keccak_sponge_context sponge_ctx;
+    size_t block_size;  /* block size in bytes */
+}
+mbedtls_shake_context;
 
 #else  /* MBEDTLS_SHA3_ALT */
 #include "sha3_alt.h"
@@ -212,6 +271,129 @@ int mbedtls_sha3( const unsigned char* input,
  * \return         0 if successful, or 1 if the test failed
  */
 int mbedtls_sha3_self_test( int verbose );
+
+/**
+ * \brief          Initialize a SHAKE context.
+ *
+ * \param ctx      The SHAKE context to initialize.
+ */
+void mbedtls_shake_init( mbedtls_shake_context *ctx );
+
+/**
+ * \brief          Clear a SHAKE context.
+ *
+ * \param ctx      The SHAKE context to be clear.
+ */
+void mbedtls_shake_free( mbedtls_shake_context *ctx );
+
+/**
+ * \brief          Clone (the state of) a SHAKE context
+ *
+ * \param dst      The destination context.
+ * \param src      The context to clone.
+ */
+void mbedtls_shake_clone( mbedtls_shake_context *dst,
+                          const mbedtls_shake_context *src );
+
+/**
+ * \brief          Start a SHAKE calculation.
+ *
+ * \param ctx      The SHAKE context to set up.
+ * \param type     The SHAKE variant to select (SHAKE128 or SHAKE256).
+ *
+ * \retval 0       Success.
+ * \retval #MBEDTLS_ERR_SHA3_BAD_INPUT_DATA
+ *                 \p ctx is \c NULL,
+ *                 or \p type is invalid,
+ *                 or this function was called without a prior call to
+ *                 mbedtls_shake_init() or after calling
+ *                 mbedtls_shake_update() or mbedtls_shake_process() or
+ *                 mbedtls_shake_ouput(),
+ */
+int mbedtls_shake_starts( mbedtls_shake_context *ctx,
+                          mbedtls_shake_type_t type );
+
+/**
+ * \brief          Feed a buffer into an ongoing SHAKE calculation.
+ *
+ * \param ctx      The SHAKE context.
+ * \param input    The buffer to process.
+ * \param size     The number of bytes to process from \p input.
+ *
+ * \retval 0       Success.
+ * \retval #MBEDTLS_ERR_SHA3_BAD_INPUT_DATA
+ *                 \p ctx is \c NULL,
+ *                 or mbedtls_shake_starts() has not been called previously,
+ *                 or mbedtls_shake_output() has been called on \p ctx.
+ */
+int mbedtls_shake_update( mbedtls_shake_context *ctx,
+                          const unsigned char* input,
+                          size_t size );
+
+/**
+ * \brief          Generate output bytes from a SHAKE calculation.
+ *
+ *                 This function can be called multiple times to generate an
+ *                 arbitrary-length output.
+ *
+ * \param ctx      The SHAKE context.
+ * \param output   Pointer to the buffer to where the output bytes are written.
+ * \param olen     The number of bytes to output.
+ *
+ * \retval 0       Success.
+ * \retval #MBEDTLS_ERR_SHA3_BAD_INPUT_DATA
+ *                 \p ctx or \p output is \c NULL,
+ *                 or mbedtls_shake_starts() has not been called previously,
+ */
+int mbedtls_shake_output( mbedtls_shake_context *ctx,
+                          unsigned char* output,
+                          size_t olen );
+
+/**
+ * \brief          Process a data block with SHAKE. For internal use only.
+ *
+ * \param ctx      The SHAKE context.
+ * \param input    The buffer containing bytes to process. The size of this
+ *                 buffer is:
+ *                 - 168 bytes for SHAKE128.
+ *                 - 136 bytes for SHAKE256.
+ *
+ * \retval 0       Success.
+ * \retval #MBEDTLS_ERR_SHA3_BAD_INPUT_DATA
+ *                 \p ctx or \p output is \c NULL.
+ *                 or mbedtls_shake_starts() has not been called previously,
+ *                 or mbedtls_shake_output() has been called on \p ctx.
+ */
+int mbedtls_shake_process( mbedtls_shake_context *ctx,
+                           const unsigned char* input );
+
+/**
+ * \brief          Generate SHAKE output from some input bytes.
+ *
+ * \param input    The buffer to process.
+ * \param ilen     The length (in bytes) of the input buffer.
+ * \param type     The SHAKE variant to calculate (SHAKE128 or SHAKE256).
+ * \param output   Pointer to the buffer to where the output data is written.
+ * \param olen     The number of output bytes to generate and write to
+ *                 \p output.
+ *
+ * \retval 0       Success.
+ * \retval #MBEDTLS_ERR_SHA3_BAD_INPUT_DATA
+ *                 \p ctx or \p output is \c NULL,
+ *                 or \c type is invalid.
+ */
+int mbedtls_shake( const unsigned char* input,
+                   size_t ilen,
+                   mbedtls_shake_type_t type,
+                   unsigned char* output,
+                   size_t olen );
+
+/**
+ * \brief          Checkup routine
+ *
+ * \return         0 if successful, or 1 if the test failed
+ */
+int mbedtls_shake_self_test( int verbose );
 
 #ifdef __cplusplus
 }
