+++ library/Makefile
@@ -74,7 +74,6 @@ OBJS_CRYPTO=	aes.o		aesni.o		arc4.o		\
 		ecp_curves.o	entropy.o	entropy_poll.o	\
 		error.o		gcm.o		havege.o	\
 		hkdf.o		hmac_drbg.o			\
-		keccak_sponge.o	keccakf.o			\
 		md.o		md2.o				\
 		md4.o		md5.o		md_wrap.o	\
 		memory_buffer_alloc.o		nist_kw.o	\
@@ -84,7 +83,7 @@ OBJS_CRYPTO=	aes.o		aesni.o		arc4.o		\
 		platform.o	platform_util.o	poly1305.o	\
 		ripemd160.o	rsa_internal.o	rsa.o  		\
 		sha1.o		sha256.o			\
-		sha3.o		sha512.o	shake.o		\
+		sha3.o		sha512.o			\
 		threading.o	timing.o	version.o	\
 		version_features.o		xtea.o
