+++ library/error.c
@@ -121,14 +121,6 @@
 #include "mbedtls/hmac_drbg.h"
 #endif
 
-#if defined(MBEDTLS_KECCAK_SPONGE_C)
-#include "mbedtls/keccak_sponge.h"
-#endif
-
-#if defined(MBEDTLS_KECCAKF_C)
-#include "mbedtls/keccakf.h"
-#endif
-
 #if defined(MBEDTLS_MD_C)
 #include "mbedtls/md.h"
 #endif
@@ -199,10 +191,6 @@
 
 #if defined(MBEDTLS_SHA512_C)
 #include "mbedtls/sha512.h"
-#endif
-
-#if defined(MBEDTLS_SHAKE_C)
-#include "mbedtls/shake.h"
 #endif
 
 #if defined(MBEDTLS_SSL_TLS_C)
@@ -783,20 +771,6 @@ void mbedtls_strerror( int ret, char *buf, size_t buflen )
         mbedtls_snprintf( buf, buflen, "HMAC_DRBG - The entropy source failed" );
 #endif /* MBEDTLS_HMAC_DRBG_C */
 
-#if defined(MBEDTLS_KECCAK_SPONGE_C)
-    if( use_ret == -(MBEDTLS_ERR_KECCAK_SPONGE_BAD_INPUT_DATA) )
-        mbedtls_snprintf( buf, buflen, "KECCAK_SPONGE - Bad input parameters to function" );
-    if( use_ret == -(MBEDTLS_ERR_KECCAK_SPONGE_NOT_SETUP) )
-        mbedtls_snprintf( buf, buflen, "KECCAK_SPONGE - mbedtls_keccak_sponge_starts has not been called" );
-    if( use_ret == -(MBEDTLS_ERR_KECCAK_SPONGE_BAD_STATE) )
-        mbedtls_snprintf( buf, buflen, "KECCAK_SPONGE - The requested operation cannot be performed with the current context state" );
-#endif /* MBEDTLS_KECCAK_SPONGE_C */
-
-#if defined(MBEDTLS_KECCAKF_C)
-    if( use_ret == -(MBEDTLS_ERR_KECCAKF_BAD_INPUT_DATA) )
-        mbedtls_snprintf( buf, buflen, "KECCAKF - Invalid input parameter(s)" );
-#endif /* MBEDTLS_KECCAKF_C */
-
 #if defined(MBEDTLS_MD2_C)
     if( use_ret == -(MBEDTLS_ERR_MD2_HW_ACCEL_FAILED) )
         mbedtls_snprintf( buf, buflen, "MD2 - MD2 hardware accelerator failed" );
@@ -890,15 +864,6 @@ void mbedtls_strerror( int ret, char *buf, size_t buflen )
     if( use_ret == -(MBEDTLS_ERR_SHA512_HW_ACCEL_FAILED) )
         mbedtls_snprintf( buf, buflen, "SHA512 - SHA-512 hardware accelerator failed" );
 #endif /* MBEDTLS_SHA512_C */
-
-#if defined(MBEDTLS_SHAKE_C)
-    if( use_ret == -(MBEDTLS_ERR_SHAKE_BAD_INPUT_DATA) )
-        mbedtls_snprintf( buf, buflen, "SHAKE - Invalid input parameter(s)" );
-    if( use_ret == -(MBEDTLS_ERR_SHAKE_BAD_NOT_STARTED) )
-        mbedtls_snprintf( buf, buflen, "SHAKE - mbedtls_keccak_sponge_starts has not been called" );
-    if( use_ret == -(MBEDTLS_ERR_SHAKE_BAD_STATE) )
-        mbedtls_snprintf( buf, buflen, "SHAKE - Requested operation cannot be performed with the current context state" );
-#endif /* MBEDTLS_SHAKE_C */
 
 #if defined(MBEDTLS_THREADING_C)
     if( use_ret == -(MBEDTLS_ERR_THREADING_FEATURE_UNAVAILABLE) )
